完成账户管理和创建单据功能 by 梅杰

+++ src/main/java/businesslogic/financialbl/StubFinancial.java
@@ -3,6 +3,7 @@ package businesslogic.financialbl;
 import java.util.ArrayList;
 
 import vo.*;
+import businesslogic.Role;
 import businesslogic.accountbl.StubAccount;
 import businesslogic.accountbl.StubAccountList;
 import businesslogic.financialbillbl.StubCashPaymentBill;
@@ -10,42 +11,43 @@ import businesslogic.financialbillbl.StubFinancialBillList;
 import businesslogic.financialbillbl.StubPaymentBill;
 import businesslogic.financialbillbl.StubReceiptBill;
 import businesslogic.iquirybl.StubInquiry;
+import businesslogic.userbl.User;
 import businesslogicservice.financialblservice.StubFinancialBlService;
 
 public class StubFinancial implements StubFinancialBlService{
 	String name;
 	String password;
+	Role role;
 	public StubFinancial() {
-		this(null, null);
+		this(null, null, null);
 	}
 	
-	public StubFinancial(String n, String pw) {
+	public StubFinancial(String n, String pw, Role r) {
 		name = n;
 		password = pw;
-	}
-	//修改密码
-	public boolean updatePassword(String newPassword) {
-		//User u = new User();
-		//return u.updatePassword(this ,newPassword);
-		return true;
+		role = r;
 	}
 	
 	public boolean addAccount(String name) {
+		if(role != Role.FINANCIAL_MANAGER) return false;
 		StubAccountList a = new StubAccountList();
 		return a.addAccount(new StubAccount(name));	
 	}
 	
 	public boolean deleteAccount(String name) {
+		if(role != Role.FINANCIAL_MANAGER) return false;
 		StubAccountList a = new StubAccountList();
 		return a.deleteAccount(new StubAccount(name));
 	}
 	
 	public AccountVO findAccount(String name) {
+		if(role != Role.FINANCIAL_MANAGER) return null;
 		StubAccountList a = new StubAccountList();
 		return a.findAccount(new StubAccount(name));
 	}
 	
 	public boolean updateAccount(String oldname, String newname) {
+		if(role != Role.FINANCIAL_MANAGER) return false;
 		StubAccountList a = new StubAccountList();
 		return a.updateAccount(new StubAccount(oldname),newname);
 	}
@@ -72,17 +74,20 @@ public class StubFinancial implements StubFinancialBlService{
 	}
 	
 	public boolean creatReceipt(ReceiptVO rv) {
+		rv.setOperator(role);
 		StubFinancialBillList financialList = new StubFinancialBillList();
 		return financialList.creatReceiptBill(rv);
 	}
 	
 	public boolean creatPayment(PaymentVO pv){
+		pv.setOperator(role);
 		StubFinancialBillList financialList = new StubFinancialBillList();
 		return financialList.creatPaymentBill(pv);
 	
 	}
 	
 	public boolean creatCashPayment(CashPaymentVO cpv) {
+		cpv.setOperator(role);
 		StubFinancialBillList financialList = new StubFinancialBillList();
 		return financialList.creatCashPaymentBill(cpv);
